implemented pop cmd
